requirements.txt
filelock==3.4.0
PyYAML==6.0
